requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cheroot==8.5.1
CherryPy==18.6.0
cryptography==3.3.1
decorator==4.4.2
et-xmlfile==1.0.1
idna==2.10
jaraco.classes==3.1.0
jaraco.collections==3.1.0
jaraco.functools==3.1.0
jaraco.text==3.4.0
jdcal==1.4.1
more-itertools==8.6.0
openpyxl==3.0.6
portend==2.7.0
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
requests==2.25.1
six==1.15.0
tempora==4.0.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
zc.lockfile==2.0
